flattened project structure - no plans for supporting additional plcs, so removed unnecessay nesting separated some modules into smaller, more focused ones refactoring and renaming simpler driver import: `from pycomm import CLXDriver`
